requirements.txt
asttokens==2.2.1
backcall==0.2.0
bleach==6.0.0
blinker==1.6.2
build==0.10.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.0
coverage==7.2.7
cx-Oracle==8.3.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
docutils==0.20.1
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
fonttools==4.40.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipython==8.14.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
keyring==24.0.1
kiwisolver==1.4.4
llvmlite==0.39.1
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ML-Monitoring @ http://${USER_SONAR}:${PSS_SONAR}@sonar.netechgroup.local:8081/repository/ml-monitoring/packages/ml-monitoring/1.0.4/ML_Monitoring-1.0.4-py3-none-any.whl
mlanomalygate @ http://${USER_SONAR}:${PSS_SONAR}@sonar.netechgroup.local:8081/repository/ml-anomaly-gate/packages/mlanomalygate/1.0.1/mlanomalygate-1.0.1-py3-none-any.whl
more-itertools==9.1.0
numba==0.56.4
numpy==1.23.5
packaging==23.1
pandas==2.0.2
parameterized==0.9.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.7.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.15.1
PyMySQL==1.0.3
pyodbc==4.0.39
pyparsing==3.1.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.1
readme-renderer==40.0
requests>=2.25.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
SQLAlchemy==1.4.39
stack-data==0.6.2
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
twine==4.0.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
xgboost==1.7.6
xmldiff==2.6.3
zipp==3.15.0
et-xmlfile==1.1.0
openpyxl==3.1.2
httpretty~=1.0.5
setuptools~=67.8.0
imblearn~=0.0
paramiko~=3.3.1
sshtunnel~=0.4.0